requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pyspellchecker>=0.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0